+++ bench/bvh/bench_gpu_bvh.mojo
@@ -511,7 +511,7 @@ def _download_full_hit_checksum(
     d_hits_f32: DeviceBuffer[DType.float32],
     ray_count: Int,
 ) raises -> Tuple[Float64, UInt32]:
-    var checksum = Float64(0.0)
+    var checksum = 0.0
     var hit_count = UInt32(0)
     with d_hits_f32.map_to_host() as h:
         for i in range(ray_count):
@@ -528,7 +528,7 @@ def _download_reduced_hit_t(
     d_partial_sums: DeviceBuffer[DType.float64],
     d_partial_counts: DeviceBuffer[DType.uint32],
 ) raises -> Tuple[Float64, UInt32]:
-    var checksum = Float64(0.0)
+    var checksum = 0.0
     var hit_count = UInt32(0)
     with d_partial_sums.map_to_host() as sums:
         for i in range(GPU_REDUCE_THREADS):
@@ -654,7 +654,7 @@ def _benchmark_direct_uploaded_rays(
     var best_kernel_ns = Int(INF_NS)
     var best_download_ns = Int(INF_NS)
     var best_frame_ns = Int(INF_NS)
-    var checksum = Float64(0.0)
+    var checksum = 0.0
     var hit_count = UInt32(0)
 
     for _ in range(repeats):
@@ -737,7 +737,7 @@ def _benchmark_camera_full_download(
     var best_kernel_ns = Int(INF_NS)
     var best_download_ns = Int(INF_NS)
     var best_frame_ns = Int(INF_NS)
-    var checksum = Float64(0.0)
+    var checksum = 0.0
     var hit_count = UInt32(0)
 
     for _ in range(repeats):
@@ -803,7 +803,7 @@ def _benchmark_primary_reduce(
     var best_reduce_ns = Int(INF_NS)
     var best_download_ns = Int(INF_NS)
     var best_frame_ns = Int(INF_NS)
-    var checksum = Float64(0.0)
+    var checksum = 0.0
     var hit_count = UInt32(0)
 
     for _ in range(repeats):
